+++ src/main/java/com/api/library/domain/BookRepository.java
@@ -1,4 +1,4 @@
-package com.api.library;
+package com.api.library.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
