+++ src/main/java/com/api/library/domain/CirculationType.java
@@ -1,4 +1,4 @@
-package com.api.library;
+package com.api.library.domain;
 
 public enum CirculationType {
     CIRCULATING, RESTRICTED
