+++ src/test/java/com/api/library/NewBookControllerTest.java
@@ -1,5 +1,6 @@
 package com.api.library;
 
+import com.api.library.usecases.registernewbook.NewBookRequest;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
